HalfEdgeDataStructure R6 Class

HalfEdgeDataStructure R6 Class

HalfEdgeDataStructure R6 Class

Public fields

BIG

A constant used to encode a half-edge as an integer rotate rotate normal vector onto axis plane

If you have a plane, you have a normal vector and an origin. Let's call your plane's normal vector the new z axis. You can generate the new y axis by crossing the old x axis with the new z axis (your plane's normal). Generate the new x axis by crossing the new z with the new y. Make all your new axis vectors into unit vectors (length 1). For every point you have, create a vector that goes from your new origin to the point (vector subtraction of point - plane_origin). Just dot with the new x and new y unit vectors and you get a pair (x,y).

Active bindings

faces

A list of faces of the input hull.

vertices

A list of vertices of the input hull.

half_edges

A list of half_edges of the input hull.

he_index

An index of the half-edges by pairs of vertex ids (as a matrix).

he_face_index

An index of the half-edges by pairs of face ids (as a matrix).

cut

Requires a spanning tree of the hull in adjacency list form.

fold

Requires a spanning tree of the dual of hull in adjacency list form.

Methods

Public methods

HalfEdgeDataStructure$rotate()

HalfEdgeDataStructure$get_angle()

HalfEdgeDataStructure$adjmat_to_edgelist()

HalfEdgeDataStructure$new()

HalfEdgeDataStructure$clone()

Method rotate()

Usage

HalfEdgeDataStructure$rotate(points, normal, around)

Arguments

points

The points to rotate

normal

The normal to a face

around

The center of the face

Returns

The rotated points

get_angle

Method get_angle()

Usage

HalfEdgeDataStructure$get_angle(begin, finish)

Arguments

begin

A beginning point (x,y vector)

finish

An ending point (x,y vector,

Returns

an angle in radians from the x-axis

adjmat_to_edgelist

Convert an adjacency matrix to an edgelist

Method adjmat_to_edgelist()

Usage

HalfEdgeDataStructure$adjmat_to_edgelist(adjmat)

Arguments

adjmat

An adjacency matrix for a graph

Returns

The corresponding edgelist for the same graph

initialize

Initialize a half-edge data structure using information in a "hull" data structure.

Method new()

Usage

HalfEdgeDataStructure$new(hull)

Arguments

hull

A data structure which is compatible with the output of the cxhull::cxhull() function.

Returns

An initalized half-edge data structure

Examples

tet <- list(
  vertices = list(
    list(id = 1, point = c(1,1,1)),
    list(id = 2, point = c(1,-1,-1)),
    list(id = 3, point = c(-1,1,-1)),
    list(id = 4, point = c(-1,-1,1))
  ),
  ridges = list(
    list(id = 1, vertices = c(1,3)), list(id = 2, vertices = c(2,3)),
    list(id = 3, vertices = c(1,2)), list(id = 4, vertices = c(3,4)),
    list(id = 5, vertices = c(1,4)), list(id = 6, vertices = c(2,4))
  ),
  facets = list(
    list(id = 1, normal = c(sqrt(3)/3,sqrt(3)/3,-sqrt(3)/3), center = c(1/3,1/3,-1/3),
         orientation = -1, vertices = c(2,1,3)),
    list(id = 2, normal = c(-sqrt(3)/3,sqrt(3)/3,sqrt(3)/3), center = c(-1/3,1/3,1/3),
         orientation = 1, vertices = c(4,1,3)),
    list(id = 3, normal = c(-sqrt(3)/3,-sqrt(3)/3,-sqrt(3)/3), center = c(-1/3,-1/3,-1/3),
         orientation = -1, vertices = c(4,2,3)),
    list(id = 4, normal = c(sqrt(3)/3,-sqrt(3)/3,sqrt(3)/3), center = c(1/3,-1/3,1/3),
         orientation = 1, vertices = c(4,2,1))
  )
)
he_ds <- HalfEdgeDataStructure$new(tet)

Method clone()

The objects of this class are cloneable with this method.

Usage

HalfEdgeDataStructure$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `HalfEdgeDataStructure$new`
## ------------------------------------------------

tet <- list(
  vertices = list(
    list(id = 1, point = c(1,1,1)),
    list(id = 2, point = c(1,-1,-1)),
    list(id = 3, point = c(-1,1,-1)),
    list(id = 4, point = c(-1,-1,1))
  ),
  ridges = list(
    list(id = 1, vertices = c(1,3)), list(id = 2, vertices = c(2,3)),
    list(id = 3, vertices = c(1,2)), list(id = 4, vertices = c(3,4)),
    list(id = 5, vertices = c(1,4)), list(id = 6, vertices = c(2,4))
  ),
  facets = list(
    list(id = 1, normal = c(sqrt(3)/3,sqrt(3)/3,-sqrt(3)/3), center = c(1/3,1/3,-1/3),
         orientation = -1, vertices = c(2,1,3)),
    list(id = 2, normal = c(-sqrt(3)/3,sqrt(3)/3,sqrt(3)/3), center = c(-1/3,1/3,1/3),
         orientation = 1, vertices = c(4,1,3)),
    list(id = 3, normal = c(-sqrt(3)/3,-sqrt(3)/3,-sqrt(3)/3), center = c(-1/3,-1/3,-1/3),
         orientation = -1, vertices = c(4,2,3)),
    list(id = 4, normal = c(sqrt(3)/3,-sqrt(3)/3,sqrt(3)/3), center = c(1/3,-1/3,1/3),
         orientation = 1, vertices = c(4,2,1))
  )
)
he_ds <- HalfEdgeDataStructure$new(tet)