hull_plus_g6 Treat the half-edge data structure as an arithmetic operator on the indices of the spanning trees

hull_plus_g6 Treat the half-edge data structure as an arithmetic operator on the indices of the spanning trees

Usage

hull_plus_g6(g6, hull)

Arguments

g6

A spanning tree in g6 encoding

hull

A hull structure for a polyhedron

Value

Vector of 4 adjacency matrices in g6 encoding. Only 2 of them will be trees, and one of those trees will be identical to the input tree.